feat: add generation to SignupForm
