+++ types/Generation.ts
@@ -1,6 +1,7 @@
 import { User } from "./User"
 
 export type Generation = {
+  id: number
   order: number
   leader?: User
 }
